api/requirements.txt
bidict==0.22.1
blinker==1.7.0
boto3==1.33.5
botocore==1.33.5
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
python-engineio==4.8.0
python-socketio==5.10.0
s3transfer==0.8.2
simple-websocket==1.0.0
six==1.16.0
urllib3==2.0.7
Werkzeug==3.0.1
wsproto==1.2.0
